+++ src/App.jsx
@@ -1,22 +1,13 @@
-// src/App.jsx
-
 import React from 'react';
+import AddTodoForm from './AddTodoForm';
+import TodoList from './TodoList';
 
 function App() {
-  const todoList = [
-    { id: 1, title: 'Complete assignment 1' },
-    { id: 2, title: 'Complete assignment 2' },
-    { id: 3, title: 'Complete assignment 3' },
-  ];
-  
   return (
     <div>
-      <h1>Lista de tareas Pendientes</h1>
-      <ul>
-        {todoList.map(item => (
-          <li key={item.id}>{item.title}</li>
-        ))}
-      </ul>
+      <h1>To Do List</h1>
+      <AddTodoForm/>
+      <TodoList/>
     </div>
   );
 }
